projects/ngcourse/meeting-pro-g/cloudrun/adk/requirements.txt
google-adk
google-cloud-aiplatform>=1.59.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-cloud-bigquery==3.14.0
python-json-logger==2.0.7
